requirements.txt
asgiref==3.8.1
boto3==1.35.63
botocore==1.35.63
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
pip==22.2.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.3
setuptools==63.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
